pyproject.toml
[tool.poetry]
name = "waternlights"
version = "0.1.0"
description = "Site para a venda de ilustrações e publicações de um blog"
authors = ["Igor Taconi <igortaconi1@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.5"
flask = "^1.1.2"
dynaconf = "^3.1.4"
python-dotenv = "^0.15.0"
gunicorn = "^20.0.4"
gevent = "^21.1.2"
peewee = "^3.14.4"
flask_simplelogin = "^0.0.7"
wtf-peewee = "^3.0.2"
Flask-Admin = "^1.5.8"
python-slugify = "^5.0.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
Flask-DebugToolbar = "^0.11.0"
ipdb = "^0.13.6"
isort = "^5.7.0"

[tool.black]
line_length = 79
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
line_length = 79
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
